requirements.txt
annotated-types==0.7.0
blinker==1.9.0
cachetools==5.5.2
certifi==2025.8.3
charset-normalizer==3.4.3
click==8.2.1
contourpy==1.3.3
cycler==0.12.1
-e git+https://github.com/agedcrow/flask-expected-value.git@c74244aa4a2c709d75bd2a5c087bd7e063de7727#egg=expected_value
Flask==3.1.2
Flask-WTF==1.2.2
fonttools==4.59.2
google-auth==2.40.3
google-auth-oauthlib==1.2.2
gspread==6.2.1
gunicorn==23.0.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
kiwisolver==1.4.9
MarkupSafe==3.0.2
matplotlib==3.10.6
numpy==2.3.3
oauthlib==3.3.1
packaging==25.0
pandas==2.3.2
pillow==11.3.0
pyasn1==0.6.1
pyasn1_modules==0.4.2
pydantic==2.11.9
pydantic_core==2.33.2
pyparsing==3.2.4
python-dateutil==2.9.0.post0
pytz==2025.2
requests==2.32.5
requests-oauthlib==2.0.0
rsa==4.9.1
setuptools==80.9.0
six==1.17.0
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
urllib3==2.5.0
Werkzeug==3.1.3
WTForms==3.2.1
